+++ docs/release_notes.rst
@@ -3,7 +3,7 @@ Release Notes
 
 v0.1.3
 ------
-* Added kmatchMixin for tests
+* Added KmatchTestMixin for tests
 
 v0.1.2
 ------